+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 Version 2.9.6.2
 --------------------------------------------------------
 - App Menu icon can be hidden from the ML4W Settings App
+- Setup script optimize the keybindings if french keyboard layout (fr) is selected
 
 Version 2.9.6.1
 --------------------------------------------------------